refactor(orchestration): move handleSuggestionClick to facade layer
Move handleSuggestionClick from App.jsx to orchestration layer:
- Create memberSelectionOrchestrator.js with
  handleSuggestionClickOrchestrated()
- Replace App.jsx implementation with thin wrapper
- Export from orchestration/index.js facade
- Add export smoke test

This is a verbatim move with no logic changes. Function body is
preserved exactly, only closure variables are passed via deps
(~128 lines of member selection and validation logic).

Dependencies: currentGroup, 8 setters, backend, fetchFrequentPartners,
isPlayerAlreadyPlaying, guardAddPlayerEarly, getCourtData,
getAvailableCourts, showAlertMessage, Tennis global.
