+++ src/registration/orchestration/index.js
@@ -14,6 +14,6 @@ export { success, failure, wrapAsync } from './helpers/resultNormalizer.js';
 // Orchestrators will be added here as they are moved:
 export { changeCourtOrchestrated } from './courtChangeOrchestrator.js';
 export { resetFormOrchestrated, applyInactivityTimeoutOrchestrated } from './resetOrchestrator.js';
-// export { handleSuggestionClickOrchestrated, handleAddPlayerSuggestionClickOrchestrated } from './memberSelectionOrchestrator.js';
+export { handleSuggestionClickOrchestrated } from './memberSelectionOrchestrator.js';
 // export { sendGroupToWaitlistOrchestrated } from './waitlistOrchestrator.js';
 // export { assignCourtToGroupOrchestrated } from './assignCourtOrchestrator.js';
